Modify code + pseudocode for single round function
revert a previous design choice to make user_choice and computer_choice
global. Previously I chose this so that later in the program could have
access to these values, but failed to implement explicity that the
function to run a single round will take them as parameters.

In addition, include specific log messages for single round outcome and
current score in the pseudocode for the single round function that
wasn't accounted for when creating the outcome variable to reflect the
current code